+++ presentation/src/main/java/com/pomonyang/mohanyang/presentation/service/PomodoroTimer.kt
@@ -1,4 +1,4 @@
-package com.pomonyang.mohanyang.presentation.screen.pomodoro.service
+package com.pomonyang.mohanyang.presentation.service
 
 internal interface PomodoroTimer {
     fun startTimer(maxTime: Int, eventHandler: PomodoroTimerEventHandler)
