+++ presentation/src/main/java/com/pomonyang/mohanyang/presentation/service/PomodoroTimerEventHandler.kt
@@ -1,4 +1,4 @@
-package com.pomonyang.mohanyang.presentation.screen.pomodoro.service
+package com.pomonyang.mohanyang.presentation.service
 
 internal interface PomodoroTimerEventHandler {
     fun onTimeEnd()
